+++ src/media/image-content.ts
@@ -2,6 +2,7 @@ import { promises } from "fs";
 import { fromBuffer } from "file-type";
 import { fetchFileAsBuffer } from "./utils";
 import { LLmMessageImageDataUrlContent, LLmMessageImageUrlContent } from "../providers";
+import { shallowFilterUndefined } from "../shared";
 
 interface BufferImage {
   type: "buffer";
@@ -23,6 +24,8 @@ interface UrlImage {
 
 type ImageSource = BufferImage | Base64Image | UrlImage;
 
+export type ImageMetadata = Record<string, number | string | boolean | null>;
+
 /**
  * Simplifies handling of images handed to JorEl. Instances of this class can be
  * created from a variety of sources, and can be passed directly to JorEl.
@@ -30,22 +33,28 @@ type ImageSource = BufferImage | Base64Image | UrlImage;
 export class ImageContent {
   /** @internal */
   private _source: ImageSource;
+  readonly metadata?: ImageMetadata;
 
-  constructor(source: ImageSource) {
+  constructor(source: ImageSource, metadata?: ImageMetadata) {
     this._source = source;
+    this.metadata = metadata;
   }
 
   /**
    * Instantiates an image from a data URL
-   * @param dataUrl
+   * @param dataUrl - The data URL of the image
+   * @param metadata - Metadata to attach to the image
    */
-  static fromDataUrl(dataUrl: string): ImageContent {
+  static fromDataUrl(dataUrl: string, metadata: ImageMetadata = {}): ImageContent {
     const [mimeType, base64] = dataUrl.split(/[:;,]/).slice(1, 3);
-    return new ImageContent({
-      type: "base64",
-      data: base64,
-      mimeType,
-    });
+    return new ImageContent(
+      {
+        type: "base64",
+        data: base64,
+        mimeType,
+      },
+      metadata,
+    );
   }
 
   /**
@@ -53,54 +62,76 @@ export class ImageContent {
    *
    * By default, the image is downloaded, since some LLMs (like Ollama
    * and Anthropic) currently do not support URLs directly.
-   * @param url The URL of the image
-   * @param mimeType The MIME type of the image
-   * @param downloadMedia Whether to download the image or not
+   * @param url - The URL of the image
+   * @param mimeType - The MIME type of the image
+   * @param downloadMedia - Whether to download the image or not
+   * @param metadata - Metadata to attach to the image
    */
-  static async fromUrl(url: string, mimeType: string, downloadMedia: false): Promise<ImageContent>;
-  static async fromUrl(url: string, mimeType?: string, downloadMedia?: true): Promise<ImageContent>;
-  static async fromUrl(url: string, mimeType?: string, downloadMedia: boolean = true): Promise<ImageContent> {
-    if (!downloadMedia) return new ImageContent({ type: "url", url, mimeType: mimeType || "" });
+  static async fromUrl(
+    url: string,
+    mimeType: string,
+    downloadMedia: false,
+    metadata?: ImageMetadata,
+  ): Promise<ImageContent>;
+  static async fromUrl(
+    url: string,
+    mimeType?: string,
+    downloadMedia?: true,
+    metadata?: ImageMetadata,
+  ): Promise<ImageContent>;
+  static async fromUrl(
+    url: string,
+    mimeType?: string,
+    downloadMedia: boolean = true,
+    metadata?: ImageMetadata,
+  ): Promise<ImageContent> {
+    if (!downloadMedia) return new ImageContent({ type: "url", url, mimeType: mimeType || "" }, metadata);
     const buffer = await fetchFileAsBuffer(url, mimeType);
-    return new ImageContent({
-      type: "buffer",
-      buffer: buffer.buffer,
-      mimeType: buffer.mimeType,
-    });
+    return new ImageContent(
+      {
+        type: "buffer",
+        buffer: buffer.buffer,
+        mimeType: buffer.mimeType,
+      },
+      metadata,
+    );
   }
 
   /**
    * Instantiates an array of images from an array of URLs
-   * @param urls Array of URLs
+   * @param urls - Array of URLs
+   * @param metadata - Metadata to attach to the images
    */
-  static async fromUrls(urls: string[]): Promise<ImageContent[]> {
-    return Promise.all(urls.map((url) => ImageContent.fromUrl(url)));
+  static async fromUrls(urls: string[], metadata?: ImageMetadata): Promise<ImageContent[]> {
+    return Promise.all(urls.map((url) => ImageContent.fromUrl(url, undefined, undefined, metadata)));
   }
 
   /**
    * Instantiates an image from an existing buffer
-   * @param buffer Buffer containing the image data
-   * @param mimeType (optional) MIME type of the image
-   * @throws Error if the MIME type cannot be detected
+   * @param buffer - Buffer containing the image data
+   * @param mimeType - (optional) MIME type of the image
+   * @param metadata - Metadata to attach to the image
+   * @throws Error - Error if the MIME type cannot be detected
    */
-  static async fromBuffer(buffer: Buffer, mimeType?: string): Promise<ImageContent> {
+  static async fromBuffer(buffer: Buffer, mimeType?: string, metadata?: ImageMetadata): Promise<ImageContent> {
     const type = mimeType || (await fromBuffer(buffer))?.mime;
     if (!type) {
       throw new Error("Unsupported image type. Unable to detect MIME type from the buffer.");
     }
-    return new ImageContent({ type: "buffer", buffer, mimeType: type });
+    return new ImageContent({ type: "buffer", buffer, mimeType: type }, metadata);
   }
 
   /**
    * Instantiates an image from a file path
-   * @param filePath Path to the image file
-   * @param mimeType (optional) MIME type of the image
-   * @throws Error if the file cannot be read
+   * @param filePath - Path to the image file
+   * @param mimeType - (optional) MIME type of the image
+   * @param metadata - Metadata to attach to the image
+   * @throws Error - Error if the file cannot be read
    */
-  static async fromFile(filePath: string, mimeType?: string): Promise<ImageContent> {
+  static async fromFile(filePath: string, mimeType?: string, metadata?: ImageMetadata): Promise<ImageContent> {
     try {
       const buffer = await promises.readFile(filePath);
-      return ImageContent.fromBuffer(buffer, mimeType);
+      return ImageContent.fromBuffer(buffer, mimeType, metadata);
     } catch (e: unknown) {
       const message = e instanceof Error ? e.message : typeof e === "string" ? e : "Unknown error";
       throw new Error(`Failed to read file: ${filePath}, error: ${message}`);
@@ -109,11 +140,12 @@ export class ImageContent {
 
   /**
    * Instantiates an array of images from an array of file paths
-   * @param files Array of file paths
-   * @param mimeType (optional) MIME type of the images
+   * @param files - Array of file paths
+   * @param mimeType - (optional) MIME type of the images
+   * @param metadata - Metadata to attach to the images
    */
-  static async fromFiles(files: string[], mimeType?: string): Promise<ImageContent[]> {
-    return Promise.all(files.map((file) => ImageContent.fromFile(file, mimeType)));
+  static async fromFiles(files: string[], mimeType?: string, metadata?: ImageMetadata): Promise<ImageContent[]> {
+    return Promise.all(files.map((file) => ImageContent.fromFile(file, mimeType, metadata)));
   }
 
   /**
@@ -174,7 +206,7 @@ export class ImageContent {
 
   /**
    * Returns the image as a message content
-   * @param downloadUrls Whether to download the image and return a data URL
+   * @param downloadUrls - Whether to download the image and return a data URL
    * @returns The image as a message content
    */
   async toMessageContent(downloadUrls?: false): Promise<LLmMessageImageDataUrlContent | LLmMessageImageUrlContent>;
@@ -183,14 +215,15 @@ export class ImageContent {
     downloadUrls: boolean = false,
   ): Promise<LLmMessageImageUrlContent | LLmMessageImageDataUrlContent> {
     if (this._source.type === "url" && !downloadUrls) {
-      return { type: "imageUrl", url: this._source.url };
+      return { type: "imageUrl", url: this._source.url, metadata: this.metadata };
     }
     const { data, mimeType } = await this.toBase64();
-    return {
+    return shallowFilterUndefined({
       type: "imageData",
       data: `data:${mimeType};base64,${data}`,
       mimeType,
-    };
+      metadata: this.metadata,
+    });
   }
 
   /**