+++ src/providers/google-vertex-ai/convert-llm-message.ts
@@ -1,6 +1,5 @@
-import { CoreLlmMessage } from "../../providers";
+import { LlmMessage, LlmUserMessageContent } from "../../providers";
 import { Content, Part } from "@google-cloud/vertexai";
-import { ImageContent } from "../../media";
 import { getBase64PartFromDataUrl } from "../../media/utils";
 
 const textContentToParts = (text: string): Part[] => {
@@ -11,57 +10,43 @@ const textContentToParts = (text: string): Part[] => {
   ];
 };
 
-const mixedContentToParts = async (content: CoreLlmMessage["content"]): Promise<Part[]> => {
-  if (typeof content === "string") {
-    return textContentToParts(content);
-  } else if (Array.isArray(content)) {
-    const parts: Part[] = [];
-    for (const _contentEntry of content) {
-      if (_contentEntry instanceof ImageContent) {
-        parts.push({
-          inlineData: await _contentEntry.toBase64(),
-        });
-      } else if (typeof _contentEntry === "string") {
-        parts.push({
-          text: _contentEntry,
-        });
-      } else if (_contentEntry.type === "text") {
-        parts.push({
-          text: _contentEntry.text,
-        });
-      } else if (_contentEntry.type === "imageData") {
-        if (!_contentEntry.mimeType) {
-          throw new Error(`Missing MIME type`);
-        }
-        parts.push({
-          inlineData: {
-            data: getBase64PartFromDataUrl(_contentEntry.data),
-            mimeType: _contentEntry.mimeType,
-          },
-        });
-      } else if (_contentEntry.type === "imageUrl") {
-        if (!_contentEntry.mimeType) {
-          throw new Error(`Missing MIME type`);
-        }
-        parts.push({
-          fileData: {
-            fileUri: _contentEntry.url,
-            mimeType: _contentEntry.mimeType,
-          },
-        });
-      } else {
-        throw new Error(`Unsupported content type`);
+const mixedContentToParts = async (content: LlmUserMessageContent[]): Promise<Part[]> => {
+  const parts: Part[] = [];
+  for (const _contentEntry of content) {
+    if (_contentEntry.type === "text") {
+      parts.push({
+        text: _contentEntry.text,
+      });
+    } else if (_contentEntry.type === "imageData") {
+      if (!_contentEntry.mimeType) {
+        throw new Error(`Missing MIME type`);
       }
+      parts.push({
+        inlineData: {
+          data: getBase64PartFromDataUrl(_contentEntry.data),
+          mimeType: _contentEntry.mimeType,
+        },
+      });
+    } else if (_contentEntry.type === "imageUrl") {
+      if (!_contentEntry.mimeType) {
+        throw new Error(`Missing MIME type`);
+      }
+      parts.push({
+        fileData: {
+          fileUri: _contentEntry.url,
+          mimeType: _contentEntry.mimeType,
+        },
+      });
+    } else {
+      throw new Error(`Unsupported content type`);
     }
-    return parts;
-  } else {
-    throw new Error(`Unsupported content type`);
   }
+  return parts;
 };
 
 /** Converts unified LLM messages to Vertex AI's messages (Content) */
 export const convertLlmMessagesToVertexAiMessages = async (
-  messages: CoreLlmMessage[],
+  messages: LlmMessage[],
 ): Promise<{ systemMessage: string | undefined; chatMessages: Content[] }> => {
   // 1. Extract system messages and join them together
   const systemMessages = messages.filter((m) => m.role === "system");