+++ src/providers/google-vertex-ai/google-vertex-ai-provider.ts
@@ -13,7 +13,7 @@ import {
 } from "@google-cloud/vertexai";
 import { FunctionDeclaration, FunctionDeclarationSchema } from "@google-cloud/vertexai/src/types/content";
 import {
-  CoreLlmMessage,
+  LlmMessage,
   generateAssistantMessage,
   LlmCoreProvider,
   LlmGenerationConfig,
@@ -104,7 +104,7 @@ export class GoogleVertexAiProvider implements LlmCoreProvider {
 
   async generateResponse(
     model: string,
-    messages: CoreLlmMessage[],
+    messages: LlmMessage[],
     config: LlmGenerationConfig = {},
   ): Promise<LlmResponse> {
     const start = Date.now();
@@ -226,7 +226,7 @@ export class GoogleVertexAiProvider implements LlmCoreProvider {
 
   async *generateResponseStream(
     model: string,
-    messages: CoreLlmMessage[],
+    messages: LlmMessage[],
     config: Omit<LlmGenerationConfig, "tools" | "toolChoice"> = {},
   ): AsyncGenerator<LlmStreamResponseChunk | LlmStreamResponse, void, unknown> {
     const start = Date.now();