+++ src/providers/mistral/mistral-provider.ts
@@ -1,7 +1,7 @@
 import { Mistral } from "@mistralai/mistralai";
 import { firstEntry, generateUniqueId, MaybeUndefined } from "../../shared";
 import {
-  CoreLlmMessage,
+  LlmMessage,
   generateAssistantMessage,
   LlmCoreProvider,
   LlmGenerationConfig,
@@ -42,7 +42,7 @@ export class MistralProvider implements LlmCoreProvider {
 
   async generateResponse(
     model: string,
-    messages: CoreLlmMessage[],
+    messages: LlmMessage[],
     config: LlmGenerationConfig = {},
   ): Promise<LlmResponse> {
     const start = Date.now();
@@ -120,7 +120,7 @@ export class MistralProvider implements LlmCoreProvider {
 
   async *generateResponseStream(
     model: string,
-    messages: CoreLlmMessage[],
+    messages: LlmMessage[],
     config: LlmGenerationConfig = {},
   ): AsyncGenerator<LlmStreamResponseChunk | LlmStreamResponse | LlmStreamResponseWithToolCalls, void, unknown> {
     const start = Date.now();