+++ src/providers/open-ai/open-ai-provider.ts
@@ -1,7 +1,7 @@
 import { OpenAI } from "openai";
 import { firstEntry, generateUniqueId, MaybeUndefined } from "../../shared";
 import {
-  CoreLlmMessage,
+  LlmMessage,
   generateAssistantMessage,
   LlmCoreProvider,
   LlmGenerationConfig,
@@ -44,7 +44,7 @@ export class OpenAIProvider implements LlmCoreProvider {
 
   async generateResponse(
     model: string,
-    messages: CoreLlmMessage[],
+    messages: LlmMessage[],
     config: LlmGenerationConfig = {},
   ): Promise<LlmResponse> {
     const start = Date.now();
@@ -105,7 +105,7 @@ export class OpenAIProvider implements LlmCoreProvider {
 
   async *generateResponseStream(
     model: string,
-    messages: CoreLlmMessage[],
+    messages: LlmMessage[],
     config: LlmGenerationConfig = {},
   ): AsyncGenerator<LlmStreamResponseChunk | LlmStreamResponse | LlmStreamResponseWithToolCalls, void, unknown> {
     const start = Date.now();